# PERF: Optimize EntityFilter.cs