Hi @WirthLukas, I've tried to make some performance optimizations to this repo. Following are the benchmark results before and after the changes:

**Benchmarks before changes:**
|                  Method |     Mean |     Error |    StdDev |    StdErr |      Min |       Q1 |   Median |       Q3 |      Max |      Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------ |---------:|----------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|----------:|-------:|------:|------:|----------:|
|  TestingClassComponents | 1.912 us | 0.0220 us | 0.0184 us | 0.0051 us | 1.887 us | 1.898 us | 1.908 us | 1.920 us | 1.952 us | 523,109.6 | 1.1787 |     - |     - |   2.41 KB |
| TestingStructComponents | 1.818 us | 0.0197 us | 0.0164 us | 0.0046 us | 1.792 us | 1.812 us | 1.817 us | 1.829 us | 1.854 us | 550,044.6 | 1.1768 |     - |     - |   2.41 KB |


**Benchmarks after changes:**
|                  Method |     Mean |     Error |    StdDev |    StdErr |      Min |       Q1 |   Median |       Q3 |      Max |      Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------ |---------:|----------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|----------:|-------:|------:|------:|----------:|
|  TestingClassComponents | 1.885 us | 0.1161 us | 0.3313 us | 0.0342 us | 1.490 us | 1.621 us | 1.822 us | 2.027 us | 2.747 us | 530,546.8 | 0.7801 |     - |     - |   1.59 KB |
| TestingStructComponents | 1.720 us | 0.1033 us | 0.3012 us | 0.0304 us | 1.387 us | 1.449 us | 1.671 us | 1.875 us | 2.531 us | 581,298.9 | 0.7801 |     - |     - |   1.59 KB |


As you can see the allocations have gone down. I've also made sure that all the unit tests pass following these changes as well. Could you please confirm whether or not you think my changes are valid?

Thank you!